requirements.txt
websockets
aiohttp
pyyaml
coinbase-advanced-py
